requirements.txt
# FastAPI framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Async support
aiofiles>=24.0.0

# PostgreSQL async driver
asyncpg>=0.29.0
aiosqlite>=0.20.0

# HTTP client
httpx>=0.27.0

# Environment variables
python-dotenv>=1.0.0

# Logging
python-json-logger>=2.0.7

# File upload
python-multipart>=0.0.12

# Type hints (using newer versions for Python 3.13 support)
pydantic>=2.10.0

# YAML parsing
PyYAML>=6.0.1

# JWT authentication
python-jose==3.5.0

# Development dependencies (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov>=5.0.0
